chore(dags): hoist shared chunk_ranges into posthog/dags/common
The marketing and web dimensional precompute warming jobs each defined a
character-identical chunk_ranges helper. Move it next to
check_for_concurrent_runs in posthog/dags/common and import it from both, so
the chunking algorithm has a single home.
