+++ posthog/dags/common/__init__.py
@@ -1,9 +1,16 @@
-from .common import check_for_concurrent_runs, dagster_tags, settings_with_log_comment, skip_if_already_running
+from .common import (
+    check_for_concurrent_runs,
+    chunk_ranges,
+    dagster_tags,
+    settings_with_log_comment,
+    skip_if_already_running,
+)
 from .owners import JobOwners
 
 __all__ = [
     "JobOwners",
     "check_for_concurrent_runs",
+    "chunk_ranges",
     "dagster_tags",
     "settings_with_log_comment",
     "skip_if_already_running",
