+++ posthog/dags/common/common.py
@@ -1,5 +1,6 @@
 from collections.abc import Callable
 from contextlib import suppress
+from datetime import datetime, timedelta
 from functools import wraps
 from typing import Optional
 
@@ -100,3 +101,20 @@ def skip_if_already_running(fn: Callable) -> Callable:
         return fn(context)
 
     return wrapper
+
+
+def chunk_ranges(start: datetime, end: datetime, chunk_days: int) -> list[tuple[datetime, datetime]]:
+    """Split [start, end) into <=chunk_days sub-windows, newest first.
+
+    Newest-first so recent data (which carries the shortest TTL and is requested most) is
+    refreshed before older history is backfilled. Shared by the dag warming jobs that drive
+    the lazy-computation framework one bounded chunk at a time.
+    """
+    chunks: list[tuple[datetime, datetime]] = []
+    cur_end = end
+    step = timedelta(days=max(chunk_days, 1))
+    while cur_end > start:
+        cur_start = max(start, cur_end - step)
+        chunks.append((cur_start, cur_end))
+        cur_end = cur_start
+    return chunks
