+++ products/marketing_analytics/dags/marketing_precompute.py
@@ -31,7 +31,7 @@ from prometheus_client import Counter
 
 from posthog.clickhouse.client.execute import KillSwitchLevel, get_kill_switch_level
 from posthog.cloud_utils import is_cloud
-from posthog.dags.common import JobOwners, check_for_concurrent_runs
+from posthog.dags.common import JobOwners, check_for_concurrent_runs, chunk_ranges
 from posthog.models import Team
 from posthog.settings import TEST
 
@@ -94,21 +94,6 @@ def get_selected_team_ids() -> list[int]:
     if raw is None:
         return list(DEFAULT_ROLLOUT_TEAM_IDS) if is_cloud() else []
     return [int(part.strip()) for part in raw.split(",") if part.strip().isdigit()]
-
-
-def chunk_ranges(start: datetime, end: datetime, chunk_days: int) -> list[tuple[datetime, datetime]]:
-    """Split [start, end) into <=chunk_days sub-windows, newest first.
-
-    Newest-first so recent data (shortest TTL, requested most) is refreshed before older history.
-    """
-    chunks = []
-    cur_end = end
-    step = timedelta(days=max(chunk_days, 1))
-    while cur_end > start:
-        cur_start = max(start, cur_end - step)
-        chunks.append((cur_start, cur_end))
-        cur_end = cur_start
-    return chunks
 
 
 def _ensure_touchpoints_for_team(
